requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
class-registry==2.1.2
Django==5.1.4
django-appconf==1.1.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
django-select2==8.4.0
django-unfold==0.48.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filters==1.3.2
frozenlist==1.5.0
geoip2==5.0.1
idna==3.10
maxminddb==2.6.3
multidict==6.1.0
pillow==11.1.0
propcache==0.3.0
PyJWT==2.10.1
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
